pyproject.toml
[project]
name = "agente-ia-youtube"
version = "0.1.0"
description = "Agente de IA que evalua videos de Youtube"
authors = [
    {name = "Tomas Gonzalez", email = "tomi150301@gmail.com"}
]
license = {text = "Tomas Daniel Gonzalez"}
readme = "README.md"
requires-python = ">=3.12,<3.15" 
dependencies = [
    "django (>=5.0,<6.0.0)",
    "djangorestframework (>=3.15.0,<4.0.0)",
    "langgraph (>=0.2.0,<3.0.0)",
    "langchain-google-genai (>=2.0.0,<5.0.0)",
    "pydantic-settings (>=2.1.0,<3.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)",
    "youtube-transcript-api (>=0.6.2,<2.0.0)",
    "aiohttp (>=3.9.0,<4.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)",
    "adrf (>=0.1.6,<0.2.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ruff = ">=0.1.0,<1.0.0"
pytest = ">=7.4.0,<10.0.0"
pytest-django = ">=4.7.0,<5.0.0"
pytest-asyncio = ">=0.23.0,<2.0.0"

[build-system]
requires = ["poetry-core>=1.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

